Y 1. 添加控制地图移入目标显示手形的配置
