+++ public/extends/cemap/earth-engine/config/earth/configIndex.d.ts
@@ -17,6 +17,7 @@ export interface EarthConfig {
     // widget类名
     widgetClassName: string;
     linkId: string;
+    showPointerAtTarget: boolean;
     // =========================相机配置==================================
     defaultView?: {
         lon?: number,
