+++ public/extends/cemap/earth-engine/layer/layer.js
@@ -170,8 +170,4 @@ var Layer = (function () {
 
 })()
 
-
-
-
-
 export default Layer
